+++ discordaio/channel.py
@@ -4,6 +4,7 @@ from .base import DiscordObject
 from .user import User
 from .role import Role
 from .emoji import Emoji
+from .invite import Invite
 from typing import List
 from enum import Enum
 
@@ -89,20 +90,22 @@ class Channel(DiscordObject):
         .. versionadded:: 0.3.0
         :param msg: The message to send. If it's over 2000 chars it will be splitted.
         :param tts: Wether to send it as a tts message.
+
+        TODO: Add embed and files.
         """
         if len(msg) <= 2000:
             await self.bot.http.request_url(f'/channels/{self.id}/messages', type='POST',
-                                       data={
-                                           'content': msg,
-                                           'tts': tts
-                                       })
+                                            data={
+                                                'content': msg,
+                                                'tts': tts
+                                            })
         else:
             for x in range(len(msg) // 2000 + 1):
                 await self.bot.http.request_url(f'/channels/{self.id}/messages', type='POST',
-                                                 data={
-                                                     'content': msg[x * 2000:x * 2000 + 2000],
-                                                     'tts': tts
-                                                 })
+                                                data={
+                                                    'content': msg[x * 2000:x * 2000 + 2000],
+                                                    'tts': tts
+                                                })
 
     async def typing(self):
         """Start typing.
@@ -144,6 +147,14 @@ class Channel(DiscordObject):
             params['after'] = after
 
         res = await self.bot.http.request_url(f'/channels/{self.id}/messages', params=params)
+        return await ChannelMessage.from_api_res(res, self.bot)
+
+    async def get_message(self, message_id) -> 'ChannelMessage':
+        """Gets a specific channel message.
+
+        .. versionadded: 0.3.0
+        """
+        res = await self.bot.http.request_url(f'/channels/{self.id}/messages/{message_id}')
         return await ChannelMessage.from_api_res(res, self.bot)
 
     async def update(self) -> 'Channel':
@@ -191,6 +202,78 @@ class Channel(DiscordObject):
         """
         res = await self.bot.http.request_url(f'/channels/{self.id}')
         return await Channel.from_api_res(res, self.bot)
+
+    async def bulk_delete_messages(self, ids: List):
+        """Delete multiple messages in a single request.
+        This endpoint can only be used on guild channels and requires the MANAGE_MESSAGES permission.
+
+        This endpoint will not delete messages older than 2 weeks,
+        and will fail if any message provided is older than that.
+
+        .. versionadded:: 0.3.0
+        """
+        if not isinstance(ids, list):
+            raise TypeError("ids must be a list of id.")
+        if not (2 <= len(ids) <= 100):
+            raise ValueError("ID list length must be between 2-100")
+
+        data = {
+            "messages": ids
+        }
+        await self.bot.http.request_url(f'/channels/{self.id}/messages/bulk-delete', type='POST', data=data)
+
+    async def get_invites(self) -> List[Invite]:
+        """Returns a list of invite objects (with invite metadata) for the channel. Only usable for guild channels.
+        Requires the MANAGE_CHANNELS permission.
+
+        .. versionadded:: 0.3.0
+        """
+        res = await self.bot.http.request_url(f'/channels/{self.id}/invites')
+        return await Invite.from_api_res(res, self.bot)
+
+    async def create_invite(self, max_age: int, max_uses=0, temporary=False, unique=False) -> Invite:
+        """Create a new invite object for the channel. Only usable for guild channels.
+        Requires the CREATE_INSTANT_INVITE permission.
+
+        .. versionadded:: 0.3.0
+        """
+        res = await self.bot.http.request_url(f'/channels/{self.id}/invites', type='POST', data={
+            "max_age": max_age,
+            "max_uses": max_uses,
+            "temporary": temporary,
+            "unique": unique
+        })
+        return await Invite.from_api_res(res, self.bot)
+
+    async def delete_permission(self, overwrite_id):
+        """Delete a channel permission overwrite for a user or role in a channel.
+        Only usable for guild channels. Requires the MANAGE_ROLES permission.
+
+        .. versionadded:: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.id}/permissions/{overwrite_id}', type='DELETE')
+
+    async def get_pinned_messages(self) -> List['ChannelMessage']:
+        """Returns all pinned messages in the channel as an array of message objects.
+
+        .. versionadded:: 0.3.0
+        """
+        res = await self.bot.http.request_url(f'/channels/{self.id}/pins')
+        return await ChannelMessage.from_api_res(res, self.bot)
+
+    async def pin_message(self, message_id):
+        """Pin a message in a channel. Requires the MANAGE_MESSAGES permission.
+
+        .. versionadded:: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.id}/pins/{message_id}', type='PUT')
+
+    async def delete_pinned_message(self, message_id):
+        """Delete a pinned message. Requires the MANAGE_MESSAGES permission.
+
+        .. versionadded:: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.id}/pins/{message_id}', type='DELETE')
 
     async def _from_api_ext(self, key, value):
         if key == 'recipients':
@@ -249,12 +332,12 @@ class Reaction(DiscordObject):
     .. versionadded:: 0.2.0
 
     Attributes:
-        count (:obj:`int`): times this emoji has been used to react
-        me (:obj:`bool`): whether the current user reacted using this emoji
-        emoji (:class:`.Emoji`): emoji information
+        count: times this emoji has been used to react
+        me: whether the current user reacted using this emoji
+        emoji emoji information
     """
 
-    def __init__(self, count=None, me=False, emoji=None):
+    def __init__(self, count: int = None, me: bool = False, emoji: Emoji = None):
         self.count = count
         self.me = me
         self.emoji = emoji
@@ -478,36 +561,39 @@ class ChannelMessage(DiscordObject):
     .. versionadded:: 0.2.0
 
     Note:
-        The author object follows the structure of the :class:`.User` object, but is only a valid user in the case where the message is generated by a user or bot user.
-        If the message is generated by a :class:`.Webhook`, the author object corresponds to the webhook's id, username, and avatar.
+        The author object follows the structure of the :class:`.User` object, but is only a valid user in the case
+        where the message is generated by a user or bot user.
+        If the message is generated by a :class:`.Webhook`,
+        the author object corresponds to the webhook's id, username, and avatar.
         You can tell if a message is generated by a webhook by checking for the webhook_id on the message object.
 
     Attributes:
-        id (:obj:`int`): id of the message
-        channel_id (:obj:`int`): id of the channel the message was sent in
-        author (user): object the author of this message (not guaranteed to be a valid user, see below)
-        content (:obj:`str`): contents of the message
-        timestamp (:obj:`int`): timestamp when this message was sent
-        edited_timestamp (:obj:`int`): timestamp when this message was edited (or null if never)
-        tts (:obj:`bool`): whether this was a TTS message
-        mention_everyone (:obj:`bool`): whether this message mentions everyone
-        mentions (:obj:`list` of :class:`.User`): objects users specifically mentioned in the message
-        mention_roles (:obj:`list` of :class:`.Role`): object ids roles specifically mentioned in this message
-        attachments (:obj:`list` of :class:`.Attachment`): objects any attached files
-        embeds (:obj:`list` of :class:`.Embed`): objects any embedded content
-        reactions (:obj:`list` of :class:`.Reaction`): objects reactions to the message
-        nonce (:obj:`int`, optional): used for validating a message was sent
-        pinned (:obj:`bool`): whether this message is pinned
-        webhook_id (:obj:`int`, optional): if the message is generated by a webhook, this is the webhook's id
-        type (:obj:`int`): type of message
-        activity (MessageActivity): activity object sent with Rich Presence-related chat embeds
-        application (MessageApplication): application object sent with Rich Presence-related chat embeds
+        id: id of the message
+        channel_id: id of the channel the message was sent in
+        author: the author of this message (not guaranteed to be a valid user, see below)
+        content: contents of the message
+        timestamp: timestamp when this message was sent
+        edited_timestamp: timestamp when this message was edited (or null if never)
+        tts: whether this was a TTS message
+        mention_everyone: whether this message mentions everyone
+        mentions: objects users specifically mentioned in the message
+        mention_roles: object ids roles specifically mentioned in this message
+        attachments: objects any attached files
+        embeds: objects any embedded content
+        reactions: objects reactions to the message
+        nonce: used for validating a message was sent
+        pinned whether this message is pinned
+        webhook_id: if the message is generated by a webhook, this is the webhook's id
+        type: type of message
+        activity: activity object sent with Rich Presence-related chat embeds
+        application: application object sent with Rich Presence-related chat embeds
     """
 
-    def __init__(self, id=None, channel_id=None, author=None, content=None, timestamp=None,
-                 edited_timestamp=None, tts=False, mention_everyone=False, mentions=[],
-                 mention_roles=[], attachments=[], embeds=[], reactions=[],
-                 nonce=None, pinned=False, webhook_id=None, type=None,
+    def __init__(self, id=None, channel_id=None, author: User = None, content: str = None, timestamp: int = None,
+                 edited_timestamp: int = None, tts: bool = False, mention_everyone: bool = False,
+                 mentions: List[User] = [], mention_roles: List[Role] = [], attachments: List[Attachment] = [],
+                 embeds: List[Embed] = [], reactions: List[Reaction] = [], nonce: int = None,
+                 pinned=False, webhook_id=None, type=None,
                  activity=MessageActivity(), application=MessageApplication()):
         self.id = id
         self.channel_id = channel_id
@@ -528,6 +614,73 @@ class ChannelMessage(DiscordObject):
         self.type = type
         self.activity = activity
         self.application = application
+
+    async def delete_own_reaction(self, emoji_id):
+        """Delete a reaction the bot has made for the message"""
+        await self.delete_user_reaction('@me', emoji_id)
+
+    async def delete_user_reaction(self, user_id, emoji_id):
+        """Deletes another user's reaction. This endpoint requires the 'MANAGE_MESSAGES'
+        permission to be present on the current user.
+
+        .. versionadded: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.channel_id}/messages/{self.id}/'
+                                        f'reactions/{emoji_id}/{user_id}', type='DELETE')
+
+    async def delete_all_reactions(self):
+        """Deletes all reactions on a message. This endpoint requires the 'MANAGE_MESSAGES'
+        permission to be present on the current user.
+
+        .. versionadded: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.channel_id}/messages/{self.id}/reactions', type='DELETE')
+
+    async def get_reactions(self, emoji_id, before: int = None, after: int = None, limit: int = None) -> List[User]:
+        """Get a list of users that reacted with this emoji.
+
+        .. versionadded: 0.3.0
+        """
+        params = dict()
+
+        if before is not None:
+            params["before"] = before
+        if after is not None:
+            params["after"] = after
+        if limit is not None:
+            if limit > 100:
+                limit = 100
+            elif limit < 1:
+                limit = 1
+            params["before"] = limit
+
+        res = await self.bot.http.request_url(f'/channels/{self.channel_id}/messages/{self.id}/'
+                                              f'reactions/{emoji_id}', params=params)
+        return await User.from_api_res(res, self.bot)
+
+    async def update(self) -> 'ChannelMessage':
+        """Updates a previously sent message with the current instance properties (content).
+        You can only edit messages that have been sent by the current user. Returns a message object.
+        Fires a Message Update Gateway event.
+
+        .. versionadded: 0.3.0
+
+        TODO: Add embed support
+        """
+        data = dict()
+        data["content"] = self.content
+        res = await self.bot.http.request_url(f'/channels/{self.channel_id}/messages/{self.id}', type='PATCH',
+                                              data=data)
+        return await ChannelMessage.from_api_res(res, self.bot)
+
+    async def delete(self):
+        """Delete a message.
+        If operating on a guild channel and trying to delete a message that was not sent by the current user,
+        this endpoint requires the MANAGE_MESSAGES permission
+
+        .. versionadded: 0.3.0
+        """
+        await self.bot.http.request_url(f'/channels/{self.channel_id}/messages/{self.id}', type='DELETE')
 
     async def _from_api_ext(self, key, value):
         if key == 'author':