+++ discordaio/permissions.py
@@ -1,0 +1,6 @@
+"""
+TODO: do this, remember to do:
+channel edit permissions endpoint
+https://discordapp.com/developers/docs/topics/permissions#permissions
+
+"""